Обновление приложения

+++ adeshina/planner/routes/events.py
@@ -1,9 +1,11 @@
 from beanie import PydanticObjectId
-from fastapi import APIRouter, HTTPException, status
+from fastapi import APIRouter, HTTPException, status, Depends
 from database.connection import Database
 
 from models.events import Event, EventUpdate
 from typing import List
+from auth.authenticate import authenticate
+
 event_database = Database(Event)
 
 
@@ -30,14 +32,22 @@ async def retrieve_event(id: PydanticObjectId) -> Event:
     return event
 
 @event_router.post("/new")
-async def create_event(body: Event) -> dict:
+async def create_event(body: Event, user: str = Depends(authenticate)) -> dict:
+    body.creator = user
     await event_database.save(body)
     return {
         "message": "Event created successfully"
     }
 
 @event_router.put("/{id}", response_model=Event)
-async def update_event(id: PydanticObjectId, body: EventUpdate) -> Event:
+async def update_event(id: PydanticObjectId, body: EventUpdate,
+                       user: str = Depends(authenticate)) -> Event:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Operation not allowed"
+        )
     update_event = await event_database.update(id, body)
     if not update_event:
         raise HTTPException(
@@ -47,13 +57,21 @@ async def update_event(id: PydanticObjectId, body: EventUpdate) -> Event:
     return update_event
 
 @event_router.delete("/{id}")
-async def delete_event(id: PydanticObjectId) -> dict:
-    event = await event_database.delete(id)
+async def delete_event(id: PydanticObjectId,
+                       user: str = Depends(authenticate)) -> dict:
+    event = await event_database.get(id)
     if not event:
         raise HTTPException(
         status_code=status.HTTP_404_NOT_FOUND,
         detail="Event with supplied ID does not exist"
     )
+    if event.creator != user:
+            raise HTTPException(
+                status_code=status.HTTP_400_BAD_REQUEST,
+                detail="Operation not allowed"
+            )
+    event = await event_database.delete(id)
+    
     return {
         "message": "Event deleted successfully."
     }
